http/tests/in-app-browser-privacy/app-bound-domain.html is a constant failure on iOS https://bugs.webkit.org/show_bug.cgi?id=210344 <rdar://problem/61583925>
Reviewed by Brent Fulgham.

Tools:

This patch creates a new TestOption parameter for enabling In-App
Browser Privacy because its NSUserDefaults key is not set using the
internal or experimental keywords. It enables this flag for
WebKitTestRunner if the option is set (and disables it otherwise).

* WebKitTestRunner/TestController.cpp:
(WTR::updateTestOptionsFromTestHeader):
* WebKitTestRunner/TestOptions.h:
(WTR::TestOptions::hasSameInitializationOptions const):
* WebKitTestRunner/cocoa/TestControllerCocoa.mm:
(WTR::TestController::platformAddTestOptions const):
* WebKitTestRunner/ios/TestControllerIOS.mm:
(WTR::TestController::platformResetStateToConsistentValues):

LayoutTests:

Update comment to match changes in TestOptions.

* http/tests/in-app-browser-privacy/app-bound-domain.html:
* http/tests/in-app-browser-privacy/switch-session-on-navigation-to-app-bound-domain.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260040 268f45cc-cd09-0410-ab3c-d52691b4dbfc