REGRESSION (r259587): bterlson/eshost throws during init in strict mode https://bugs.webkit.org/show_bug.cgi?id=210470
Reviewed by Ross Kirsling.

This change makes $262.IsHTMLDDA of JSC shell a CustomValue, allowing it to be reassigned
and restoring compatibility with any version of https://github.com/bterlson/eshost.

Since putDirectCustomAccessor() is now used instead of putGetter(), scope exception assert
is no longer needed and can be safely removed, as well as JSObject::putGetter() export.

* jsc.cpp:
* runtime/JSObject.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260057 268f45cc-cd09-0410-ab3c-d52691b4dbfc