REGRESSION (r259898): Tabs occasionally go blank and get stuck that way https://bugs.webkit.org/show_bug.cgi?id=210520 <rdar://problem/61776830>
Reviewed by Simon Fraser.

* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::setAcceleratedCompositingRootLayer):
setThumbnailView calls setAcceleratedCompositingRootLayer in order to reconnect the
existing root layer. After r259898, this would cause us to go down the
process-swap path, and hide the root layer, and nothing would ever
come along and fix it.

If setAcceleratedCompositingRootLayer is called with the existing root
layer, that cannot be a process swap, so ignore it and proceed as before.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260104 268f45cc-cd09-0410-ab3c-d52691b4dbfc