REGRESSION(r260246): It broke build on MIPS32 https://bugs.webkit.org/show_bug.cgi?id=210665
Patch by Angelos Oikonomopoulos <angelos@igalia.com> on 2020-04-18
Reviewed by Aakash Jain.

The mnemonic for 'store halfword' is 'sh', not 'shv'. This appears to
be a typo in a path that was never exercised.

Exposed by r260246; riscLowerMisplacedAddresses now calls into
riscAsRegisters with an 'h' suffix, which results in a 'storeh'
instruction. The storeh is then lowered to the non-existent 'shv'.

* offlineasm/mips.rb:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260309 268f45cc-cd09-0410-ab3c-d52691b4dbfc