Oversized caret and selection rects in text fields on ganji.com and netflix.com/login https://bugs.webkit.org/show_bug.cgi?id=210622 <rdar://problem/45945636>
Reviewed by Darin Adler.

Source/WebCore:

Currently, selection and caret rects in text fields on some web pages can be excessively tall. This patch makes
a small adjustment to allow the top of the caret or selection rect to snap to the top of the inline box instead
of being at the end of the previous line, in the case where there is no previous inline box.

In the case where we compute the caret rect for an empty renderer (i.e. no children), we make an additional
tweak so that the caret rect's height is based on the computed font height instead of line height, and then we
ensure that the caret is (logically) vertically centered.

See below for more details.

Test: editing/selection/selection-and-caret-do-not-extend-to-line-height.html

* rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::positionForPointWithInlineChildren):

Specify ForHitTesting::Yes when asking for selectionTop().

* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::localCaretRectForEmptyElement):

Use FontMetric's height when computing the height of the caret rect, and then center it vertically in the
renderer.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::positionForPoint):

Specify ForHitTesting::Yes when asking for selectionTop(). See below for more information.

* rendering/RenderTextLineBoxes.cpp:
(WebCore::RenderTextLineBoxes::positionForPoint const):
* rendering/RootInlineBox.cpp:
(WebCore::RootInlineBox::selectionTop const):

When computing selectionTop(), we currently fall back to using the top of the containing RenderBlockFlow
(`blockFlow().borderAndPaddingBefore()`) in the case where there is no previous root box. However, this can lead
to selection and caret rects being taller than expected; instead, we can use the max of the `selectionTop`
(that is, the top of the line box, adjusted for annotations) and the top of the RenderBlockFlow. This has the
effect of allowing the caret and selection to visually snap to the top of a run of text, provided there is not
already a line of text that precedes it. Taking the maximum of the two values ensures that we don't
unintentionally make the selection or caret rects even larger, if the line top is above the top of the block.

Note that we also avoid shrinking the selection and caret rects when hit-testing renderers for positions and
ranges. This allows users to still click and drag to select text in the extra line-height area above a piece of
text, even if the selection is only painted over the text (and not in the region containing the line-height).
This behavior was established in the fix for webkit.org/b/14911, and is covered by the layout test
`editing/selection/inline-closest-leaf-child.html`.

* rendering/RootInlineBox.h:

Tools:

Rebaseline some expected selection and caret rects in some iOS API tests.

* TestWebKitAPI/Tests/ios/AccessibilityTestsIOS.mm:
(TestWebKitAPI::TEST):
* TestWebKitAPI/Tests/ios/AutocorrectionTestsIOS.mm:
(TEST):
* TestWebKitAPI/Tests/ios/DragAndDropTestsIOS.mm:
(TestWebKitAPI::TEST):

LayoutTests:

Rebaseline existing tests, and add a new test to verify that caret and selection rects do not extend to match
the line height in empty and non-empty editable elements, in both horizontal and vertical writing modes.

* editing/selection/ios/absolute-selection-after-scroll-expected.txt:
* editing/selection/ios/absolute-selection-after-scroll.html:
* editing/selection/ios/change-selection-after-tapping-focused-element-expected.txt:
* editing/selection/ios/change-selection-after-tapping-focused-element.html:
* editing/selection/ios/fixed-selection-after-scroll-expected.txt:
* editing/selection/ios/fixed-selection-after-scroll.html:
* editing/selection/ios/selection-after-changing-text-with-callout-menu-expected.txt:
* editing/selection/ios/selection-after-changing-text-with-callout-menu.html:
* editing/selection/ios/selection-handles-after-touch-end-expected.txt:
* editing/selection/ios/update-selection-after-overflow-scroll-expected.txt:
* editing/selection/ios/update-selection-after-overflow-scroll.html:
* editing/selection/selection-and-caret-do-not-extend-to-line-height-expected.txt: Added.
* editing/selection/selection-and-caret-do-not-extend-to-line-height.html: Added.
* fast/repaint/focus-ring-repaint-expected.txt:
* platform/ios/editing/selection/ios/selection-handles-after-touch-end-expected.txt: Removed.

Delete this redundant test expectation.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260367 268f45cc-cd09-0410-ab3c-d52691b4dbfc