Media controls tracks menu shows "Auto" selected instead of track selected via the JS API https://bugs.webkit.org/show_bug.cgi?id=211230 <rdar://problem/62648409>
Patch by Antoine Quint <graouts@apple.com> on 2020-05-04
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/modern-media-controls/tracks-support/tracks-support-text-track-selected-via-media-api.html

We're fixing two issues with the captions menu on macOS here.

First, if a text track was marked as "showing" with the JS API, we would not show it as selected in the UI
because MediaControlsHost would report that the captionDisplayMode was "automatic" and we'd take this as
sufficient data to say that the "Automatic (Recommended)" item should be shown as selected. We now only
do this if we also don't have any text tracks set as "showing".

The second issue was when trying to select "Automatic (Recommended)" when a text track had been marked as
"showing" with the JS API. Calling `setSelectedTextTrack()` on MediaControlsHost in this case was not sufficient
because HTMLMediaElement::setSelectedTextTrack is a no-op if the automatic text track is provided but captionDisplayMode
is still set to "automatic". To address this, we first disable all text tracks before calling `setSelectedTextTrack()`.

* Modules/modern-media-controls/media/tracks-support.js:
(TracksSupport.prototype.tracksPanelIsTrackInSectionSelected):
(TracksSupport.prototype.tracksPanelSelectionDidChange):

LayoutTests:

Add a test that:

- sets an embedded text track as selected via the HTMLMediaElement TextTrack APIs,
- brings up the tracks panel,
- dumps its initial UI state, checking that the track selected via the API shows as selected in the UI,
- selects the "Automatic (Recommended)" track,
- brings up the tracks panel again,
- dumps its current UI state, checking that the "Automatic (Recommended)" shows as selected in the UI.

* media/modern-media-controls/tracks-support/tracks-support-text-track-selected-via-media-api-expected.txt: Added.
* media/modern-media-controls/tracks-support/tracks-support-text-track-selected-via-media-api.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261084 268f45cc-cd09-0410-ab3c-d52691b4dbfc