REGRESSION (JazzkonE): ResourceLoadStatisticsDatabaseStore checks 'hasHadUserInteraction' without ensuring the domain has been added to the ITP database (211388) https://bugs.webkit.org/show_bug.cgi?id=211388 <rdar://problem/62849919>
Reviewed by John Wilander.

A spurious log entry is created if we check 'hasHadUserInteraction' before inserting a row for the domain
being checked (if that domain had not previosly been observed). This patch changes the order of the two
commands to avoid this possibility. This makes the database implementation behave like the in-memory version.

* NetworkProcess/Classifier/ResourceLoadStatisticsDatabaseStore.cpp:
(WebKit::ResourceLoadStatisticsDatabaseStore::logUserInteraction):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261115 268f45cc-cd09-0410-ab3c-d52691b4dbfc