Crash under _LSSetApplicationInformationItem() https://bugs.webkit.org/show_bug.cgi?id=211478 <rdar://problem/62201314>
Reviewed by Alex Christensen.

Given the crashes, I suspect it is not actually safe to call _LSSetApplicationInformationItem()
from a non main-thread like it was done in r238289. We still run the code asynchronously to
address the issue that r238289 was trying to fix but we now call _LSSetApplicationInformationItem()
on the main thread.

* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::updateProcessName):
(WebKit::WebProcess::updateActivePages):
(WebKit::setProcessNameQueue): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261206 268f45cc-cd09-0410-ab3c-d52691b4dbfc