Cut and paste from Google Doc to Notes in several (non-Latin) languages doesn't work https://bugs.webkit.org/show_bug.cgi?id=211498 <rdar://problem/56675345>
Reviewed by Darin Adler.

Source/WebCore:

When copying text in Google Docs, the page uses `DataTransfer.setData` to write text/html data to the system
pasteboard. This markup string includes a meta tag with `charset="utf-8"`, indicating that the HTML string that
was copied should be interpreted as UTF-8 data.

However, before we write this data to the system pasteboard, we first sanitize it by loading it in a separate
page, and then build the final sanitized markup string to write by iterating over only visible content in the
main document of this page. Importantly, this last step skips over the meta element containing the charset.

Later, when pasting in Notes or TextEdit, both apps use `-[NSAttributedString initWithData:...:]` to convert the
HTML data on the pasteboard into an NSAttributedString. This takes the NSPasteboard's HTML data (a blob of
`NSData`) and synchronously loads it in a new legacy WebKit view by calling `-[WebFrame
loadData:MIMEType:textEncodingName:baseURL:]`, passing in `nil` as the text encoding name. Since WebKit is only
given a blob of data and no particular encoding, we fall back to default Latin-1 encoding, which produces
gibberish for CJK text.

To fix this, we automatically insert a `<meta charset="utf-8">` tag when writing HTML to the pasteboard, if the
sanitized markup contains non-ASCII characters.

Test: CopyHTML.SanitizationPreservesCharacterSet

* Modules/async-clipboard/ClipboardItemBindingsDataSource.cpp:
(WebCore::ClipboardItemBindingsDataSource::ClipboardItemTypeLoader::sanitizeDataIfNeeded):

Pass in AddMetaCharsetIfNeeded::Yes.

* dom/DataTransfer.cpp:
(WebCore::DataTransfer::setDataFromItemList):

Pass in AddMetaCharsetIfNeeded::Yes here too.

* editing/cocoa/WebContentReaderCocoa.mm:
(WebCore::sanitizeMarkupWithArchive):
(WebCore::WebContentReader::readHTML):
(WebCore::WebContentMarkupReader::readHTML):
* editing/markup.cpp:
(WebCore::sanitizeMarkup):

Add a new enum so that we only add the extra meta tag when sanitizing content that is being written to the
system pasteboard through one of the clipboard DOM APIs.

(WebCore::sanitizedMarkupForFragmentInDocument):
* editing/markup.h:

Source/WebKit:

Add a new header to allow Cocoa code to reason about UIColors and NSColors on iOS and macOS (respectively)
without requiring platform ifdefs. A followup patch will adopt this in several places in WebKit, where we
currently need ifdefs for iOS and macOS.

* Platform/cocoa/CocoaColor.h: Added.
* WebKit.xcodeproj/project.pbxproj:

Tools:

Add a test to verify that when writing markup to the clipboard via DOM API, if non-ASCII characters appear in
the written markup, they can still be converted to `NSAttributedString`s containing the expected non-Latin text.

* TestWebKitAPI/Configurations/Base.xcconfig:

Adjust header search paths so that we can import CocoaColor.h in WebKit.

* TestWebKitAPI/Tests/WebKitCocoa/CopyHTML.mm:
(readHTMLDataFromPasteboard):
(readHTMLStringFromPasteboard):
(readHTMLFromPasteboard): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261247 268f45cc-cd09-0410-ab3c-d52691b4dbfc