[WebXR] Implement simulateUserActivation() https://bugs.webkit.org/show_bug.cgi?id=211516
Reviewed by Youenn Fablet.

simulateUserActivation() is the method used by the WebXR Test API to
mock a user activation and thus let WebXR API think that the user is
behind the request that is going to be made in the function passed as
argument.

No new tests added as the imported WebXR web-platform-tests already make
use of this API.

* testing/WebXRTest.cpp:
(WebCore::WebXRTest::simulateUserActivation): Implemented.
* testing/WebXRTest.h: Added Document parameter.
* testing/WebXRTest.idl: Call with Document.
* testing/XRSimulateUserActivationFunction.h: Removed function params.
* testing/XRSimulateUserActivationFunction.idl: Ditto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261273 268f45cc-cd09-0410-ab3c-d52691b4dbfc